nova_boot_middleware::resilience

Function build_circuit_breaker_backend

pub fn build_circuit_breaker_backend(
    name: &str,
    cfg: &CircuitBreakerConfig,
    store_opt: Option<Arc<dyn ResilienceStore>>,
) -> Arc<dyn CircuitBreakerBackend>

Build a CircuitBreakerBackend from CircuitBreakerConfig and an optional distributed store.

Helper used by application startup code to create either a local or distributed backend depending on configuration.